requirements.txt
Django==5.2.4
Pillow==11.3.0
django-crispy-forms==2.4
crispy-bootstrap5==2025.6